+++ lib/karafka/web/pro/ui/views/consumers/commands/_breadcrumbs.erb
