+++ lib/karafka/web/pro/ui/views/consumers/commands/_command_details.erb
