+++ lib/karafka/web/pro/ui/views/consumers/commands/_incompatible_schema.erb
