+++ lib/karafka/web/pro/ui/views/consumers/commands/_table.erb
@@ -14,7 +14,7 @@
       <%==
         each_partial(
           @command_messages,
-          'commands/command',
+          'consumers/commands/command',
           local: :command_message
         )
       %>
