+++ lib/karafka/web/pro/ui/views/consumers/consumers/_consumer_performance.erb
