+++ lib/karafka/web/pro/ui/views/consumers/consumers/_tabs.erb
