+++ lib/karafka/web/pro/ui/views/consumers/consumers/consumer/_consumer_group.erb
@@ -1,7 +1,7 @@
 <%==
   render_each(
     consumer_group.subscription_groups,
-    'consumers/consumer/_subscription_group',
+    'consumers/consumers/consumer/_subscription_group',
     local: :subscription_group,
     locals: { consumer_group: consumer_group }
   )
