+++ lib/karafka/web/pro/ui/views/consumers/consumers/consumer/_no_subscriptions.erb
