+++ lib/karafka/web/pro/ui/views/consumers/consumers/consumer/_partition.erb
