+++ lib/karafka/web/pro/ui/views/consumers/consumers/index.erb
@@ -1,7 +1,7 @@
 <% view_title 'Consumers' %>
 
 <%== partial 'consumers/summary' %>
-<%== partial 'consumers/tabs' %>
+<%== partial 'consumers/consumers/tabs' %>
 
 <div class="col-span-12">
   <% if @processes.empty? && params.current_page <= 1 %>
