--- lib/karafka/web/pro/ui/views/consumers/details.erb
@@ -1,13 +1,0 @@
-<% view_title @process.id %>
-
-<%== partial 'consumers/consumer/commands', locals: { process: @process } %>
-
-<% if @process.status == 'stopped' %>
-  <%== partial 'consumers/consumer/stopped' %>
-<% end %>
-
-<%== partial 'consumers/consumer/metrics' %>
-
-<%== partial 'consumers/consumer/tabs' %>
-
-<pre class="code"><code class="json p-0 m-0"><%= JSON.pretty_generate(@process.to_h) %></code></pre>
