--- lib/karafka/web/pro/ui/views/consumers/subscriptions.erb
@@ -1,25 +1,0 @@
-<% view_title @process.id %>
-
-<%== partial 'consumers/consumer/commands', locals: { process: @process } %>
-
-<div class="col-span-12">
-  <% if @process.status == 'stopped' %>
-    <%== partial 'consumers/consumer/stopped' %>
-  <% end %>
-
-  <%== partial 'consumers/consumer/metrics' %>
-
-  <%== partial 'consumers/consumer/tabs' %>
-
-  <% if @process.subscribed? %>
-    <%==
-      render_each(
-        @process.consumer_groups,
-        'consumers/consumer/_consumer_group',
-        local: :consumer_group
-      )
-    %>
-  <% else %>
-    <%== partial 'consumers/consumer/no_subscriptions' %>
-  <% end %>
-</div>
