+++ spec/lib/karafka/web/pro/ui/controllers/consumers/commanding_controller_spec.rb
