+++ spec/lib/karafka/web/pro/ui/controllers/consumers/commands_controller_spec.rb
